requirements.txt
certifi==2020.12.5
chardet==3.0.4
click==7.1.2
fastapi==0.62.0
h11==0.11.0
idna==2.10
pydantic==1.7.3
pymongo==3.11.2
python-dateutil==2.8.1
requests==2.25.0
six==1.15.0
starlette==0.13.6
urllib3==1.26.2
uvicorn==0.13.1
gunicorn==20.0.4
uWSGI==2.0.19.1
